授業ごとに関連するイベントを表示
正直この修正はどうかと思う・・・
