+++ src/course/index.jsx
@@ -4,8 +4,17 @@ import {
   CardTitle, 
   CardText
 } from 'material-ui/Card';
+import {
+  Table,
+  TableBody,
+  TableHeader,
+  TableHeaderColumn,
+  TableRow,
+  TableRowColumn
+} from 'material-ui/Table';
 
 import userData from '../testdata';
+import EventList from '../eventlist';
 
 export default class Course extends React.Component {
   constructor(props) {
@@ -13,6 +22,7 @@ export default class Course extends React.Component {
     this.courses = userData.courses;
     this.weeks = userData.weeks;
     this.periods = userData.periods;
+	this.events = userData.events;
   }
 
   static title(params) {
@@ -36,6 +46,7 @@ export default class Course extends React.Component {
             return `${week} ${period}`;
           }).join(', ')}
         </CardText>
+		<EventList event={this.events.filter((a) => a.courseid.indexOf(Number(this.props.params.id)) !== -1)} />
       </Card>
     );
   }
