+++ src/oneevent/index.jsx
@@ -1,0 +1,1 @@
+//一つ一つのイベントの詳細を書く
